src/metaphysical/mathematics/core/omni_transformation.py
import numpy as np
import torch
import torch.nn as nn
from typing import Dict, List, Any, Optional, Tuple, Callable
from dataclasses import dataclass
from datetime import datetime
import logging
from qiskit import QuantumCircuit, QuantumRegister, ClassicalRegister, execute, Aer
from qiskit.algorithms import VQE, QAOA
from qiskit.optimization import QuadraticProgram
from qiskit.optimization.algorithms import MinimumEigenOptimizer
import time
import json
import yaml
from scipy import signal
import networkx as nx
from ..scalable_quantum_system import ScalableQuantumSystem

logger = logging.getLogger(__name__)

@dataclass
class TransformationState:
    """Represents the state of the Omni-Saiyan transformation system"""
    energy_density: float
    compactification_radius: float
    entanglement_entropy: float
    spacetime_metric: np.ndarray
    last_update: float
    system_status: str

class OmniTransformationSystem:
    """Implements the Omni-Saiyan transformation framework"""
    
    def __init__(self, num_dimensions: int = 10):
        self.backend = Aer.get_backend('qasm_simulator')
        self.quantum_system = ScalableQuantumSystem(num_qubits=128)
        self.state = TransformationState(
            energy_density=0.0,
            compactification_radius=1e-18,  # Initial Planck-scale radius
            entanglement_entropy=0.0,
            spacetime_metric=np.zeros((num_dimensions, num_dimensions)),
            last_update=time.time(),
            system_status='initialized'
        )
        
    def calculate_energy_density(self, curvature: np.ndarray, field_strength: np.ndarray) -> float:
        """Calculate energy density from Einstein-Yang-Mills stress-energy tensor"""
        try:
            # Constants
            G = 6.67430e-11  # Gravitational constant
            g = 1.0  # Yang-Mills coupling
            
            # Calculate Ricci tensor components
            R_mu_nu = np.trace(curvature)
            R = np.sum(np.diag(curvature))
            
            # Calculate metric tensor
            g_mu_nu = np.eye(curvature.shape[0])
            
            # Calculate Einstein tensor
            G_mu_nu = R_mu_nu - 0.5 * R * g_mu_nu
            
            # Calculate Yang-Mills term
            F_mu_nu = field_strength
            F_term = np.trace(np.dot(F_mu_nu, F_mu_nu))
            
            # Calculate energy density
            rho = (1 / (8 * np.pi * G)) * np.sum(G_mu_nu * g_mu_nu) + (1 / (4 * g**2)) * F_term
            
            return float(rho)
            
        except Exception as e:
            logger.error(f"Error calculating energy density: {str(e)}")
            raise
            
    def calculate_frequency(self, compactification_radius: float, laplacian_eigenvalue: float) -> float:
        """Calculate frequency from string vibrational modes"""
        try:
            c = 299792458  # Speed of light
            f = (c / (2 * np.pi * compactification_radius)) * np.sqrt(laplacian_eigenvalue)
            return float(f)
            
        except Exception as e:
            logger.error(f"Error calculating frequency: {str(e)}")
            raise
            
    def calculate_entanglement_entropy(self, area: float, num_microstates: int) -> float:
        """Calculate entanglement entropy across compact manifold"""
        try:
            G = 6.67430e-11  # Gravitational constant
            k_B = 1.380649e-23  # Boltzmann constant
            
            # Calculate entropy
            S = (area / (4 * G)) + np.log(num_microstates)
            return float(S * k_B)
            
        except Exception as e:
            logger.error(f"Error calculating entanglement entropy: {str(e)}")
            raise
            
    def calculate_spacetime_metric(self, mass: float, charge: float, angular_momentum: float,
                                 compact_coordinates: np.ndarray) -> np.ndarray:
        """Calculate modified Kerr-Newman metric with extra dimensions"""
        try:
            G = 6.67430e-11  # Gravitational constant
            c = 299792458  # Speed of light
            
            # Initialize metric tensor
            num_dimensions = len(compact_coordinates) + 4
            metric = np.zeros((num_dimensions, num_dimensions))
            
            # Calculate Kerr-Newman components
            r = np.linalg.norm(compact_coordinates[:3])
            theta = np.arccos(compact_coordinates[2] / r)
            a = angular_momentum / (mass * c)
            
            Sigma = r**2 + a**2 * np.cos(theta)**2
            Delta = r**2 - 2 * G * mass * r / c**2 + a**2
            
            # Set metric components
            metric[0,0] = -(1 - 2 * G * mass * r / (c**2 * Sigma))
            metric[1,1] = Sigma / Delta
            metric[2,2] = Sigma
            metric[3,3] = (r**2 + a**2 + 2 * G * mass * a**2 * r * np.sin(theta)**2 / (c**2 * Sigma)) * np.sin(theta)**2
            
            # Add compact dimensions
            for i in range(4, num_dimensions):
                metric[i,i] = 1.0
                
            return metric
            
        except Exception as e:
            logger.error(f"Error calculating spacetime metric: {str(e)}")
            raise
            
    def transform(self, initial_state: Dict[str, Any]) -> Dict[str, Any]:
        """Execute Omni-Saiyan transformation protocol"""
        try:
            # Calculate energy density
            energy_density = self.calculate_energy_density(
                initial_state['curvature'],
                initial_state['field_strength']
            )
            
            # Calculate frequency
            frequency = self.calculate_frequency(
                self.state.compactification_radius,
                initial_state['laplacian_eigenvalue']
            )
            
            # Calculate entanglement entropy
            entropy = self.calculate_entanglement_entropy(
                initial_state['area'],
                initial_state['num_microstates']
            )
            
            # Calculate spacetime metric
            metric = self.calculate_spacetime_metric(
                initial_state['mass'],
                initial_state['charge'],
                initial_state['angular_momentum'],
                initial_state['compact_coordinates']
            )
            
            # Update state
            self.state.energy_density = energy_density
            self.state.entanglement_entropy = entropy
            self.state.spacetime_metric = metric
            self.state.last_update = time.time()
            self.state.system_status = 'transformed'
            
            return {
                'energy_density': energy_density,
                'frequency': frequency,
                'entanglement_entropy': entropy,
                'spacetime_metric': metric,
                'system_status': self.state.system_status
            }
            
        except Exception as e:
            logger.error(f"Error executing transformation: {str(e)}")
            raise
            
    def get_transformation_report(self) -> Dict[str, Any]:
        """Generate comprehensive transformation report"""
        return {
            'timestamp': datetime.now(),
            'energy_density': self.state.energy_density,
            'compactification_radius': self.state.compactification_radius,
            'entanglement_entropy': self.state.entanglement_entropy,
            'spacetime_metric_shape': self.state.spacetime_metric.shape,
            'last_update': self.state.last_update,
            'system_status': self.state.system_status
        } 