requirements-test.txt
numpy>=1.21.0
scipy>=1.7.0
qiskit>=0.34.0
tensorflow>=2.8.0
transformers>=4.15.0
matplotlib>=3.4.0
pytest>=6.2.5
pytest-cov>=3.0.0
hypothesis>=6.24.0
unittest-mock>=0.2.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0
pytest-html>=3.2.0
pytest-metadata>=3.0.0
pytest-benchmark>=3.4.1
pytest-randomly>=3.12.0
coverage>=7.0.0
black>=21.5b2
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910
boto3>=1.26.0
python-dotenv>=0.19.0
pyyaml>=5.4.0
loguru>=0.5.3
requests>=2.26.0
flask>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0 